apps/azure_functions/requirements.txt
azure-core==1.34.0
azure-functions==1.23.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
